[MSHARED-690] Change package from org.apache.maven.it to org.apache.maven.shared.verifier
