+++ src/main/java/org/apache/maven/shared/verifier/util/ResourceExtractor.java
@@ -1,4 +1,4 @@
-package org.apache.maven.it.util;
+package org.apache.maven.shared.verifier.util;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
